simulator/requirements.txt
# Core dependencies
requests==2.28.1
python-dotenv==0.21.0
schedule==1.1.0
flask==2.2.2

# For image generation
pillow==9.3.0
numpy==1.23.5

# For mock hardware (not used but included for reference)
# These would be needed on actual Raspberry Pi hardware
# RPi.GPIO==0.7.1
# adafruit-circuitpython-dht==3.7.0
# adafruit-circuitpython-mcp3xxx==1.4.10 